Add basic sync (upload only, no deletion) when build button is clicked. Move and refactor upload code so it is available for builds as well as the initial bind process.
