+++ dev/src/codewind/project/Requester.ts
@@ -11,6 +11,9 @@
 
 import * as vscode from "vscode";
 import * as request from "request-promise-native";
+import * as path from "path";
+import * as fs from "fs-extra";
+import * as zlib from "zlib";
 
 import Project from "./Project";
 import ProjectCapabilities, { StartModes } from "./ProjectCapabilities";
@@ -34,9 +37,6 @@ namespace Requester {
     // By enforcing this and using these to forward all Codewind requests to the 'req' function,
     // we can inject options to abstract away required configuration like using json, handling ssl, and authentication.
 
-    export async function put(url: string | vscode.Uri, options?: request.RequestPromiseOptions): Promise<any> {
-        return req(request.put, url, options);
-    }
     async function req(method: RequestFunc, url: string | vscode.Uri, options?: request.RequestPromiseOptions): Promise<any> {
         if (url instanceof vscode.Uri) {
             url = url.toString();
@@ -162,8 +162,59 @@ namespace Requester {
 
         // return doProjectRequest(project, url, body, Requester.post, "Build");
         const buildMsg = Translator.t(STRING_NS, "build");
+        if (project.connection.remote) {
+            const localPath = MCUtil.fsPathToContainerPath(project.localPath);
+            Log.i(`Copying updated files from ${localPath} to ${project.connection.host}`);
+            await requestUploadsRecursively(project.connection, project.id, localPath, localPath);
+        } else {
+            Log.i(`Local build from local file system at ${project.localPath}`);
+        }
         await doProjectRequest(project, ProjectEndpoints.BUILD_ACTION, body, Requester.post, buildMsg);
-        // await requestValidate(project, true);
+    }
+
+    export async function requestUploadsRecursively(connection: Connection, projectId: any, inputPath: string, parentPath: string): Promise<void> {
+        Log.i(`requestUploadsRecursively for ${projectId} at ${inputPath}`);
+        const files = fs.readdirSync(inputPath);
+
+        for (const f of files) {
+            const currentPath = `${inputPath}/${f}`;
+            // Log.i("Uploading " + currentPath);
+            const stats = fs.statSync(currentPath);
+            if (stats.isFile()) {
+                try {
+                    await remoteUpload(connection, projectId, currentPath, parentPath);
+                } catch (err) {
+                    Log.d(err);
+                }
+            } else if (stats.isDirectory()) {
+                await requestUploadsRecursively(connection, projectId, currentPath, parentPath);
+            }
+        }
+    }
+
+    async function remoteUpload(connection: Connection, projectId: string, filePath: string, parentPath: string): Promise<any> {
+
+        const fileContent = JSON.stringify(fs.readFileSync(filePath, "utf-8"));
+        const strBuffer = zlib.deflateSync(fileContent);
+        const base64Compressed = strBuffer.toString("base64");
+        const relativePath = path.relative(parentPath, filePath);
+
+        const remoteBindUpload = EndpointUtil.resolveProjectEndpoint(connection, projectId, ProjectEndpoints.UPLOAD);
+        const body = {
+            directory: false,
+            path: relativePath,
+            timestamp: Date.now(),
+            msg: base64Compressed,
+        };
+
+        const remoteBindRes = await Requester.put(remoteBindUpload, {
+            json: true,
+            body: body,
+        });
+
+        Log.i("Remote upload response", remoteBindRes);
+
+        return remoteBindRes;
     }
 
     export async function requestToggleAutoBuild(project: Project): Promise<void> {
